+++ inventory/models/reservation.py
@@ -31,6 +31,17 @@ class ReservationStatus(models.TextChoices):
 
 BLOCKING_STATUSES = (ReservationStatus.RESERVED, ReservationStatus.PENDING)
 
+class ReservationGroup(models.Model):
+    user = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name="reservation_groups",
+    )
+    reference = models.CharField(max_length=32, unique=True, blank=True)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return f"{self.reference or self.pk}"
 
 class Reservation(models.Model):
     user = models.ForeignKey(
