+++ templates/emails/reservation_confirmed/reservation_confirmed.txt
