+++ templates/emails/reservation_created/reservation_created.txt
