+++ templates/emails/reservation_rejected/reservation_rejected.html
