+++ templates/emails/reservation_rejected/reservation_rejected.txt
